+++ modules/validators/src/cpp/ValidatorsInternal.hpp
@@ -76,5 +76,8 @@ mustBeA(const ArrayOf& arg, const wstringVector &classNames, int argPosition, bo
 NLSVALIDATORS_IMPEXP void
 mustBePositive(const ArrayOf& arg, int argPosition, bool asCaller = false);
 //=============================================================================
+NLSVALIDATORS_IMPEXP void
+mustBeNonnegative(const ArrayOf& arg, int argPosition, bool asCaller = false);
+//=============================================================================
 }
 //=============================================================================