+++ modules/validators/src/include/Validators.hpp
@@ -72,5 +72,8 @@ mustBeA(const ArrayOfVector& args, const wstringVector &classNames, int argPosit
 NLSVALIDATORS_IMPEXP void
 mustBePositive(const ArrayOfVector& args, int argPosition);
 //=============================================================================
+NLSVALIDATORS_IMPEXP void
+mustBeNonnegative(const ArrayOfVector& args, int argPosition);
+//=============================================================================
 }
 //=============================================================================