+++ CHANGELOG.md
@@ -7,7 +7,7 @@
   - `mustBeFile`,
   - `mustBeNonempty`,
   - `mustBeA`,
-  - `mustBePositive`,
+  - `mustBePositive`, `mustBeNonnegative`
 
 - `test_run` manages `SEQUENTIAL TEST REQUIRED` and `NATIVE_ARCHITECTURE TEST REQUIRED` tags.